+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControl.cs
@@ -1113,7 +1113,17 @@ namespace UnityEngine.InputSystem
             MarkAsStale();
 
             foreach (var inputControl in children)
+            {
                 inputControl.MarkAsStale();
+                if (inputControl is ButtonControl buttonControl)
+                {
+                    // If everything is becoming stale, update all press states so we can reevaluate
+                    buttonControl.UpdateWasPressed();
+                    #if UNITY_EDITOR
+                    buttonControl.UpdateWasPressedEditor();
+                    #endif
+                }
+            }
         }
 
         #if UNITY_EDITOR